backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
ExifRead==3.5.1
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.0.0
PyJWT==2.10.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.4